+++ trunk/OpenHoldem_Scraper_Optimization/CVersus.cpp
@@ -155,7 +155,7 @@ bool CVersus::GetCounts(void)
 								1617000, 1620675, 1623125, 1624350
 							  };
 	int betround = p_betround_calculator->betround();
-	int sym_userchair = (int) p_symbol_engine_userchair->userchair();
+	int sym_userchair = p_symbol_engine_userchair->userchair();
 
 	unsigned int	pcard[2] = {0};
 	for (int i=0; i<=1; i++)
